+++ src/main/resources/db/changelog/changeset/4.create-beer-table.sql
@@ -5,7 +5,7 @@ create table beer
 id bigserial primary key,
 name varchar(20) not null,
 description text,
-volume double precision  check (volume>0) not null,
+volume decimal(4,3)  check (volume>0) not null,
 price decimal(7,2)  check (price>0) not null,
 brand_id bigint references beer_brand (id) on delete cascade not null
 );
